Renamed PenPal to AttackMate
This is the magic commit that changes the project-name from penpal
to attackmate. Bow down reverently!
