+++ images/attackmate-schema.drawio
