+++ pyproject.toml
@@ -1,9 +1,9 @@
 [project]
-name = "penpal"
+name = "attackmate"
 authors = [
     {name = "Wolfgang Hotwagner", email = "wolfgang.hotwagner@ait.ac.at"},
 ]
-description = "PenPal is an attack orchestration tool that executes full attack-chains based on playbooks."
+description = "AttackMate is an attack orchestration tool that executes full attack-chains based on playbooks."
 readme = "README.md"
 requires-python = ">=3.7"
 keywords = ["Pentest", "Attack", "Orchestration", "Hacking", "Simulating", "Attackchain"]
@@ -20,7 +20,7 @@ dependencies = [
 dynamic = ["version"]
 
 [project.scripts]
-penpal = "penpal.__main__:main"
+attackmate = "attackmate.__main__:main"
 
 [build-system]
 requires = ["setuptools"]
@@ -28,7 +28,7 @@ build-backend = "setuptools.build_meta"
 
 # [tool.setuptools]
 # py-modules = []
-# packages = ["penpal"]
+# packages = ["attackmate"]
 # include-package-data = true
 
 [tool.setuptools.packages.find]
@@ -36,4 +36,4 @@ namespaces = true
 where = ["src"]
 
 [tool.setuptools.package-data]
-"penpal.data" = ["*.tar.gz"]
+"attackmate.data" = ["*.tar.gz"]
