+++ src/attackmate/regexexecutor.py
