+++ src/attackmate/sliversessionexecutor.py
