+++ src/attackmate/variablestore.py
